modif console route

+++ routes/users.js
@@ -239,23 +239,23 @@ router.post('/photo', async (req, res) => {
   //const photoPath = `./tmp/${uniqid()}.jpg`;
   const photoPath = path.join(process.cwd(), `${uniqid()}.jpg`);
   console.log(photoPath)
-  /*const resultMove = await req.files.photoFromFront.mv(photoPath);
-  console.log( "result move console", resultMove) */
+  const resultMove = await req.files.photoFromFront.mv(photoPath);
+  console.log( "result move console", resultMove) 
 
-  /*if (!resultMove) {
+  if (!resultMove) {
     
-    const resultCloudinary = await cloudinary.uploader.upload(photoPath);
+    /* const resultCloudinary = await cloudinary.uploader.upload(photoPath);
     console.log("result cloudinary back", resultCloudinary)
 
-    fs.unlinkSync(photoPath);
+    fs.unlinkSync(photoPath); 
 
-    res.json({ result: true, url: resultCloudinary.secure_url });
+    res.json({ result: true, url: resultCloudinary.secure_url }); */
+     res.json( {result: true, photo : photoPath })
   }
   
   else {
     res.json({ result: false, error: resultMove });
-  } */
-  res.json( {result: true, photo : photoPath })
+  } 
 
 });
